// restapi/dist/api/users/types.ts
/* eslint-disable */
// tslint:disable
/**
 * rest api
 * rest api
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UsersApi {
	postUsersCreate: (request: Api.User) => Promise<PostUsersCreateResponse>
	deleteUsersDelete: (id: string) => Promise<DeleteUsersDeleteResponse>
	getUsersGet: (id: string) => Promise<GetUsersGetResponse>
	getUsersGetAll: () => Promise<GetUsersGetAllResponse>
	putUsersUpdate: (request: Api.User) => Promise<PutUsersUpdateResponse>
}

export type PostUsersCreateResponse = PostUsersCreate200Response | PostUsersCreate400Response

export interface PostUsersCreate200Response {
	status: 200
	body: Api.User
	headers?: never
}

export interface PostUsersCreate400Response {
	status: 400
	body: Api.Message
	headers?: never
}

export type DeleteUsersDeleteResponse = DeleteUsersDelete200Response | DeleteUsersDelete400Response

export interface DeleteUsersDelete200Response {
	status: 200
	body: Api.Message
	headers?: never
}

export interface DeleteUsersDelete400Response {
	status: 400
	body: Api.Message
	headers?: never
}

export type GetUsersGetResponse = GetUsersGet200Response | GetUsersGet404Response

export interface GetUsersGet200Response {
	status: 200
	body: Api.User
	headers?: never
}

export interface GetUsersGet404Response {
	status: 404
	body: Api.Message
	headers?: never
}

export type GetUsersGetAllResponse = GetUsersGetAll200Response | GetUsersGetAll404Response

export interface GetUsersGetAll200Response {
	status: 200
	body: Api.User[]
	headers?: never
}

export interface GetUsersGetAll404Response {
	status: 404
	body: Api.Message
	headers?: never
}

export type PutUsersUpdateResponse = PutUsersUpdate200Response | PutUsersUpdate400Response

export interface PutUsersUpdate200Response {
	status: 200
	body: Api.User
	headers?: never
}

export interface PutUsersUpdate400Response {
	status: 400
	body: Api.Message
	headers?: never
}